Moved ShardByShard and ReedSolomon into core module
